+++ src/Symfony/Component/Form/CHANGELOG.md
@@ -17,6 +17,7 @@ CHANGELOG
  * removed the `$scale` argument of the `IntegerToLocalizedStringTransformer`
  * removed `TemplatingExtension` and `TemplatingRendererEngine` classes, use Twig instead
  * passing a null message when instantiating a `Symfony\Component\Form\FormError` is not allowed
+ * removed support for using `int` or `float` as data for the `NumberType` when the `input` option is set to `string`
 
 4.4.0
 -----